Reuse PRAW mock objects across tests.
